[INTERNAL] TableDelegates: Refactor tags in JSDoc of the classes
Added the @namespace tag, because MDC delegates are no classes. No
instance can be created, not even a singleton. They are more like utils.

According to the current guidelines, if an entity is not exported via
global name, it should not be documented with a global name
(@alias sap.ui.mdc.TableDelegate), but with a module name
(@alias module:sap/ui/mdc/TableDelegate)

Change-Id: Iafe1e8283b3f6dc1380d4ba9ebc9137e36fe1f58