[INTERNAL] ColumnListItem: center align for popinDisplay=Inline
Columns having popinDisplay=Inline are now center aligned when they move to the popin row. These column are now rendered with display:flex and align-items:center.
The visualization of the colon rendered next to the popin header has also changed in a way the in case of long column header text, the colon wraps with the header text and not as a separate flex element. For this the display: inline-flex is replaced with display: inline. The popin header label now works in a similar way to the Simple form when text is wrapped.
BCP: 2180408075
JIRA: CPOUIFTEAMB-1332

Change-Id: I7b07adf374a551ae20ed3a4870b9c8608a844597
(cherry picked from commit 00543a242bdeccdfa6f95778da6e36765e807d14)
CR-Id: 002075125800004544232021