[INTERNAL] sap.m.IllustratedMessage: Added new illustrations
Added the latest illustrations.

Additionally a text fallback private logic is introduced for the default
texts of the title and the description.
Now if we have an Illustration which ends with "_v*", we use the default
text from the Illustration without the "_v*" in the end.
Also a new map is introduced where some of the new illustrations texts
correspond to some of the old ones.

JIRA: BGSOFUIPIRIN-5490
Change-Id: I57062ac51b4f25160929daa9b938d82a5f76c0e9