[INTERNAL] core.dnd: Get the valid DropInfo from the DOM/control tree
 - Before this change, when we receive the dragenter event from the
UIArea we were traversing up on the control tree to find out the valid
DropInfo object of the target element.
 - In some cases, the control tree does not reflect the DOM tree. A
rendered control, that exists in the DOM tree, might not be the part of
the control tree. In composition cases the control tree might be
different than the rendering tree.
 - Since UI5 Drag & Drop concept based on the native DOM APIs, we will
traverse up on the DOM tree and get the valid DropInfo object from the
control for which belong to that DOM node.

Change-Id: I64d8b67a96545c32c3247567af6193cd8c2c81cc