[FIX] sap.m.SuggestionsPopover: Highligting in list removed from popover
The Suggestions Popover used to attach event delegates used for list highlighting,
but this was needed only for the Input/MultiInput, since the Combo/MultiComboBox
highlight the list by themselves.
This resuts in an incorrect highlighting of items on open in the ComboBox.
The problem does not occur on higher versions, since this logic was removed from the SP.

Change-Id: Ia964e1125e84ee2edba5c7856f668cc23f5057f2
BCP: 2180415299
CR-Id: 002075125800004423372021