[INTERNAL] sap.m.Checkbox: Remove for attribute from label
The "for" attribute on the checkbox's label element is unnecessary as the
label is referenced using aria-labelledby. What is more is that when validating the html using W3C html validtor
it detects that the label's for attribute is only allowed to reference an real input elements.

Impact: Improve the screen reader support and assure html validators pass with success.

Risk: Low risk with regards to checkbox's label announcment.

Change-Id: Idfb7448940eaa5eec7dda5a93d1d5d13ccf0b112
CR-Id: 002075125800000276472022
BCP: 002075129500008513672021
(cherry picked from commit 4084ec350e667d1b929cad391d68b4de646cebd7)