[INTERNAL] ColumnMenu: Fix re-rendering after invalidation
If the user of the menu does not add it to the control tree, the menu
will not be re-rendered when already open, because it has no connection
to the UIArea - invalidated controls are "collected" by the UIArea,
which triggers re-rendering in certain intervals.

To fix the issue, the menu is added to the static UIArea (it ends up to
be a child of the "normal" UIArea).

Change-Id: I15eba266e6578d4b4575a5c5e47853d79a9e442d