[INTERNAL] StandardListItem: fix updating displaySize property of avatar
This is a follow-up to the comment mentioned in the change I93080adbb268fce0ec210e6ea758bb8071e673bc. In case of growing, there could be cases that the child control (Avatar), could invalidate the List control, when new items are added. To prevent this, we make use of the ManagedObject's setProperty API for setting the displaySize of the Avatar and prevent invalidation.

Change-Id: I4679604b6db72422ca4bf42d467c1b6d03bba245