[FIX] sap.m.ObjectHeader: initial icon alt property could be overridden
Problem description:
In case of an incorrect sap.m.ObjectHeader's icon URI,
a broken image is displayed instead, due to screen reader
announcment improvement in relation to the same.
The alternative text alogside the broken image display can
also be helpful for sighted users. Nevertheless there are
cases when this is unwanted behavior and there is currently
no way for the application developers to use the old control behavior.

Solution:
Application developers can now override the initial "iconAlt"
property with an empty string value. That way the internally
used sap.m.Image control instance wouldn't be displayed when
its URI is incorrect.

BCP: 002075129500008606402021
Change-Id: I4e42f524ec69c75a03eb18044d80168cc1145a66