[INTERNAL] mdc.Table: Restrict DTA/RTA actions
Table.designtime.js file is updated with suitable properties that can be allowed for DTA. The internals of the table cannot be changed by RTA and hence it has been disabled.
By default, any property/aggregation added to the control is enabled for the designtime. Hence to overcome this issue and have more control over the enablement of the property/aggregation, the Table's designtime ignores all properties/aggregations for adaptation and should be explicitly added to the allowed list to enable adaptation.
JIRA: CPOUIFTEAMB-1920

Change-Id: I602640d46752be4d78fb5dcca901739fcc71fabb