[INTERNAL] sap.m.MultiInput: Improve startup performance of the control
The MultiInput creates internally a Tokenizer and adjusts its dimensions,
so it would fit and would position properly the internal tokens.
Those adjustments are not necessary when there are no tokens as the
Tokenizer is just a plain container and would not affect the layout without
Tokens inside.

JIRA: BGSOFUIRILA-2886
Change-Id: I034f9596e35babf7fe0318b57d9526d00f223dc5