[INTERNAL] ListBase: restrict RTA action when item aggregation is bound
RTA enables removing/adding of list items. This does not work as expects, when the items aggregation is bound. For example, when a single item is removed, then all the items are removed since RTA removes the template. Similar behavior when 1 item is added to the list, RTA adds all. To handle this correctly, the designtime document of the ListBase and the Table controls now disables item level actions, when the items aggregation is bound.
BCP: 2280001034

Change-Id: I02b7096a2f9eee5efdae4da2f146f5e5e2569d79