[INTERNAL][FIX] sap.ui.fl: remove optional parameter in variant handling
the parameter 'changeInstance' in the VariantManagementState.getInitialChanges
and VariantManagementState.getControlChangesForVariant should always be true,
but in 1 case it was not set.
Since now all consumers use this parameter with true it was removed completely.

Scenario: variant should be already applied when the app is loaded, then
the xml change applying created a new change to apply. So then there were
2 change instances for the same change

BCP: 2280016550
Change-Id: I1f039ec392ec49743c9279ce8a3410a67c3e9306