[FEATURE] mvc.XMLView: declarative usage of aggregation, association and event handler
The XMLTemplateProcessor is refactored to re-use the same code for
processing a View's or Fragment's root node that is used for normal
child nodes (control nodes).

This enables all kinds of settings for the View itself (e.g.
aggregations, associations and event handlers).

To further increase the reuse of the control-processing code in the case
of XHTML nested into UI5 controls, the nested XML is temporarily wrapped
in an XMLView node. Similarly, fragments that don't start with a
FragmentDefinition are wrapped as well.

For compatibility with the old implementation, any settings for a view
that are declared in the XML are applied after the settings given in the
View factory call.

The improved handling of View settings theoretically allows to bind the
content aggregation. As this would conflict with the current rendering
approach of XMLViews (preserving XHTML content and re-rendering nested
UI5 controls at the places defined in the original XML), declaring a
binding for the content aggregation of a View causes a runtime error.

Fixes: https://github.com/SAP/openui5/issues/1355
Fixes: https://github.com/SAP/openui5/issues/2662
Fixes: https://github.com/SAP/openui5/issues/3431
JIRA: CPOUI5FRAMEWORK-337
Change-Id: Iba466447e42e98d213a124f44d3ac4e64289f887