[FIX] sap.ui.core.format.NumberFormat: Stricter parsing for grouping separators
In the user input, the decimal separator is often confused with the grouping separator.
The parsing logic ignored the grouping separator, which led to an unwanted parsed
number. This change adjusts the parsing, such that it is stricter when parsing the
grouping separator, if exactly one is present.

The examples below are based on the German locale standard settings (de):
  var oOptions = {
    groupingSeparator : ".",
    decimalSeparator : ",",
    groupingSize : 3
  };
  var oNumberFormat = NumberFormat.getFloatInstance(oOptions);

Cases:
* There is exactly one grouping separator present, no decimal separator,
  and the grouping separator at the most right grouping position is wrong:

  oNumberFormat.parse("1.2"); // NaN (before: 12)
  oNumberFormat.parse("1.234567"); // NaN (before: 1234567)

  Assumption: The user most likely meant to use a decimal separator.

* The number with the grouping separator has an invalid syntax:

  oNumberFormat.parse("0.123"); // NaN (before: 123)

  Assumption: The number with a zero in front of the grouping separator is
  an indicator that the user meant a decimal separator instead.

Combines:
677e54f19cef72bd6b58249e0c7f19c0951115ee
4824d5381e1aafd50b26969535d93916454e2519
4223f5cd97847100f79f1d5deaf008d5b6f9f58a

CR-Id: 002075125800000754842022
Change-Id: Ib79518f28c980f8e885409c4c2cd813c6aa49c1a
BCP: 2180381028
JIRA: CPOUI5FOUNDATION-435
(cherry picked from commit 677e54f19cef72bd6b58249e0c7f19c0951115ee)