[FIX] sap.m.MessagePopover: Item bindings are now cloned correctly and with bound model
This change re-implements the correction in 0ebc878b9f6faa4a6aa9e1c676fadd927ba7f2f5
which fixes a typo in the cloning of the sap.m.MessageView items.

It was previously reverted because it revealed
a critical flaw of the cloning method - when 'cloneBindings'
flag is set to 'true' it bounds the default model to
the element and, not the 'messages model'.

Because of that, in the MessageView control
this 'messages' model binding is done explicitly before
every rendering (by the _bindToMessageModel() method).

Similarly, in this change we are now bounding
the 'messages' model every time we destroy and recreate
the MessageView items in the onBeforeRenderingPopover()
method, in order to overcome the core's cloning issue.

BCP: 2280014165
BCP: 2280016537
BCP: 2180409546
BCP: 1970404487
JIRA: BGSOFUIRILA-2594

Change-Id: I666963bf09d0b0dbd771293fc1022c8c22f34ab4