[INTERNAL] sap.ui.rta: remove possibility to disable addXML via designtime
by accident it was possible to disable the addXML developer action
via the designtime by setting the action to null or not-adaptable.
As the change type should not be changeable anyways, it is now always overwritten.
This way there is still the possibility to use addXML in combination with jsOnly.

BCP: 2280008676
Change-Id: I798bea51f32eaa1901e7da8332c4d2e563b4fc8c