[FIX] sap.m.MultiComboBox: Fix focus on item selection
There was a bug in the onAfterRenderingList() method that was trying
to focus non-visible items, causing the focus to be lost to the body and
the popover to not close on click outside of it.

Now the method take into account the visibility of the items and focuses
the one that was selected by the user interaction.

BCP: 2180426976
BCP: 002075129500000372792022
Change-Id: I231965cf09d0b0dbd771293fc1022c8c22f35eaf