[FIX] sap.uxap.ObjectPageLayout: ObjectPageState tests fixed
Since we counted on timeouts to make sure the scrolling to certain section and UI updates, related to that are accomplished there were some instability during tests runs. Now we call onScroll event earlier to make sure all the needed calculation at the moment of assertion are already made.

BCP: 2280011569
Change-Id: I11441a055739f50607707b4ff5a35f17ba3fd101