[INTERNAL] ListBase: restrict RTA action when item aggregation is bound
RTA enables removing/adding of list items. This does not work as expects, when the items aggregation is bound. For example, when a single item is removed, then all the items are removed since RTA removes the template. Similar behavior when 1 item is added to the list, RTA adds all. To handle this correctly, the designtime document of the ListBase and the Table controls now disables item level actions, when the items aggregation is bound.
- sap.ui.rta: remove old test 'AggregationBinding'
The test was introduced along with logic that is not there anymore.
The test relies on control implementation that is not up to us (and actually should be changed)
The logic is tested in other tests
BCP: 2280001034

CR-Id: 002075125800000441412022
Change-Id: I564664bb9e78b07a9137848f692f25d267443d0a