[INTERNAL] sap.f.DynamicPage: Fixed SemanticPage designtime qUnit in IE
Problem: A designtime qUnit test is unstable due to the use of the
'getBoundingClientRect' browser API on a DOM Node which is not a part of
the body at a certain point during testing. IE doesn't support such
scenario.

Solution: We use specific code (older revision) in the DynamicPage which
is supported by the IE browser during the problematic testing.

BCP: 2280003766
CR-Id: 002075125800000465072022
Change-Id: Ib36752c993097809fba1d0557b539be575095446