[INTERNAL] sap.m.Wizard sample: Set the step as current when invalidated
When the data in the step gets invalidated we are now setting it
as 'currentStep' in order to display the confirmation button for that
step again.

In most cases the button is labeled "Next",
but sometimes it is also used as a "Submit" button
that submits only the data in single current step.

In such case, this will enable users to submit the data only in the current step.

BCP: 2170320759
Change-Id: Ib2a1fc9f34c2bd9ee4c7ea841ce8d2097031f24a