[INTERNAL] ListBase: prevent RTA changes to items and its content when bound
In an earlier fix: I02b7096a2f9eee5efdae4da2f146f5e5e2569d79, we prevent RTA changes to the items aggregation when it is bound with "actions: null". But its turn out that, it is not enough since, the item's inner controls were still adaptable which lead to similar template removal issue as seen in the previous commit.
With this change, we now prevent actions like remove and rename on the items aggregation and its content.
We cannot complete disable RTA by saying "actions: null", since certain action like "settings" must be allowed for adaptable. For example, for the case of the SmartLink and SmartTable.
BCP: 2280001034

Change-Id: Ib8621a7f78b8d61114b41eea869b75ab8a763486