[INTERNAL] sap.m.DynamicDateRange: visual tests stabilised
Each time when a new test object was included all the generated images
were failing, because they strictly depended on the container which was shot.

Now only some of the images will be failing and it would be easier
to decide if there is a regression in the control or the differense
in the images are caused only because of the addition of the new test object.

Change-Id: Idf471702a4df4eea95f5656b856d7d78169fd42f