[INTERNAL][FIX] sap.ui.fl: Create all personalization changes before applying them
When multiple changes are added at once via the ControlPersonalizationWriteAPI,
make sure to first create and add all of them before starting the apply process.
This is necessary to make sure that change handlers can wait for all changes
and batch them together, for example to only send a single backend request
when multiple columns are added to a MDC table.

Change-Id: I847c77efb19eb614d92e856280eec090bb3a71ed
BCP: 2280026809