[INTERNAL][FIX] Table: Remove text selection when SelectAll checkbox is pressed
On double click on the SelectAll checkbox the text from the column headers gets selected which skips
the next click event. Disabling the selection on the entire column header causes other issues that
have been reported and fixed earlier. Therefore the only acceptable solution is to remove the
selection in the onclick event handler of the Pointer extension.

BCP: 2180435908
Change-Id: Icb686b58118e9326357adebc1d74c5d63126c5d5