[FIX] sap.m.SinglePlanningCalendar: appointmentSelect event is properly fired
Problem description:
The 'appointmentSelect' event isn't fired when a user taps over a full
day appointment. An internal logic is relying on that there will always
be a 'data-sap-ui-related' attribute for each appointment element.

Solution:
It is taken into consideration that the full day appointments aren't
rendered with a 'data-sap-ui-related' attribute.

BCP: 2280035754
Change-Id: Idd019db40537527b24f1168becfc452ed7f65262