[FIX] sap.m.ObjectNumber: role group is removed for non interactive control instnaces
Problem description:
Role group was added for non interactive control instances in relation
to a request from a stakeholder to reference labels to the same.
This results in redundant screen reader announcments when a sap.m.ObjectNumber
is used inside a sap.m.Table cell.

Solution:
The group role is now removed as per accessibility specification.
A non interactive sap.m.ObjectNumber control instances are displayed
as plain text, which isn't supposed to be referenced with
additional labels and descriptions.

BCP: 2180187431
Change-Id: I6826d16a907e7ff5c0765c8b00c5decf87276917