[FIX] sap.m.SinglePlanningCalendarGrid: proper year set in grid cells
* there were a wrong formatter patters set for formatters used when
constructing some ARIA texts and attributes data-sap-start-date and
data-sap-end-date of the grid cells. These formatters used 'YYYY'
instead 'yyyy' which leads to setting incorrect year on some cells
near year end (when first days of the new year are in the last week
of previous one), because 'YYYY' means 'year of week'.

* now proper format patterns (with 'yyyy') are set.

Change-Id: Ia6308ad6b4ea1d4e3075f710ce4b31b2fc7b0ebc
BCP: 002075129500000709492022