[INTERNAL][FIX] GrowingEnablement: Context change with ODataV4
If the binding context is changed, and the table is bound to a V4 model,
a "change" event is fired at the beginning with an empty diff object.
The GrowingEnablement does not update the items now, but waits for the
data. If the request fails, "change" is not fired again, so the list
keeps showing the old items, which is not the case for other controls
and if growing is not enabled.

The GrowingEnablement now listens to the "dataReceived" event and
destroys the items in case of a failing request after a context change.

Snippix id for testing:
V2: #741762
V4: #868747

Change-Id: Ib8c8c3bc9b89d114d17a8c3774fc9b0e87bca04a