[INTERNAL][FIX] sap.ui.rta: Invalid element on reveal
In some cases (e.g. when undoing a split operation) an element can be
made invalid (e.g. destroyed) asynchronously while the checks for
the additionalElementsPlugin are still running. This would cause the
invalid element to show up on the add dialogue or prevent the context
menu from showing up.

Change-Id: I3d8b9fdb422c03b24931ebe024a90c2f6be2a791
BCP: 2280034056