[INTERNAL] sap.ui.mdc - Export capabilities added
The sap.ui.mdc.TableDelegate now contains a function
that returns the export capabilities. This allows the table
to activate specific export types with specific options.

Any subclass of this Delegate that overwrites this
function has to explicitly define the type XLSX which
is the default type. This enables the application or FE
to define either PDF or XLSX or both to be active.

Change-Id: Idb7f068dc4b216492ff8deab557c5b606a585d34