[INTERNAL] LoaderExtensions: Remove setTimeout for resource data conversion
Removes the setTimeout call from the async path of the loadResource() function.
The async loadResource API already returns a Promise.
The original code including the timeout might in certain situations contribute
to a longer setTimout(0) chain that gets throttled by the browser.

The issue was brought up through an application that creates multiple
nested XML fragments. Each XML is preloaded, yet the parsing  was done
in a setTimeout, which created part of said timeout chain.

Change-Id: Ib3921877c2b49b9cd2495608d0b45001fb441214
BCP: 2180353921