[INTERNAL] sap.ui.unified.Calendar: year and year range pickers work properly
Issue: when the year to be displayed is less than half of the
year range shown from the YearPicker, the latter was not working.
There were invalid years shown, which were causing errors and invalid
keyboard navigation. Same is true for year range picker.

Solution: now wrong years are not shown and the inner calculations are adjusted.

Change-Id: Ia53c03afeaa089545fb8206a552dc0c44d9aca5f