[INTERNAL] sap.uxap.ObjectPageLayout: prevent pixel rounding issue
Problem: In _getClosestScrolledSectionBaseId we obtain the currently
closest Section/SubSection to the position, where the user has scrolled
to. However, in some cases, due to the ceil rounding of the top
positions of the SectionBases, it may happen that the current scroll
position is rounded down and it differ with just 1 pixel with the
closest SectionBase. In this case, we wrongly selected the
Section/SubSection, which is placed just above the right one.

Solution: Add 1 pixel tolerance, when comparing the top position of the
SectionBases with the current scroll position.

JIRA: BGSOFUIPIRIN-5426
Change-Id: I174c34c301f16ae6722ec0594eb4c5bcece944bb