[INTERNAL] ScrollEnablement: Break while loop if html is reached
There are case where the function getChildPosition ends in an infinite while loop
because of the $Container element is not positioned:

position: static;
or
position: fixed;

and is not returned as offsetParent.

Add condition to break the while loop if the html element is retuned or null.

Change-Id: Icdbe5eea6e7e0e987b1612aa8209ae0a2c6c2877
BCP: 002075129400000735512022
(cherry picked from commit 9e7ce87d8cc6c427a69925e02b6a59509bdb2035)
CR-Id: 002075125900002111352022