[INTERNAL][FIX] MDC Table: Merge requests in a batch in multi-table view
If there is a view with more than one table, and all tables use the
sap.ui.mdc.odata.v4.TableDelegate, then their requests will not be
merged in one batch. The delegate requires a property helper class with
sap.ui.require, which always creates a macrotask.
To keep it simple: The batch requests are sent in different macrotasks.

To avoid this issue, the property helper class is added to the module
dependencies to be already available on init.

Change-Id: I4590c820ae6d4b8085f8572608b5cce84c4c229b
BCP: 2180428223