[INTERNAL] sap.ui.core.format.DateFormat.DateFormatTimezone: timezone parameter defaulting
The timezone parameter will be defaulted if omitted, null or
empty string otherwise if the parameter is not a string a TypeError is
thrown. If the parameter is an invalid IANA timezone ID,
an error is logged, #format will return an empty string and #parse
returns null.

JIRA: CPOUI5FOUNDATION-461
Change-Id: Ifd6fa16849d90690a19a3cef34523c94c294c37a