[FEATURE] noData aggregation for custom noData visualization added
Added a 'noData' aggregation to 'ListBase' to allow for custom noData visualizations.

When provided both a 'noDataText' property and a 'noData' aggregation, the aggregation will take priority.
When provided a 'noData' aggregation with a sap.m.IllustratedMessage, a sap.m.IllustratedMessage is also used when no columns are visible.

JIRA: CPOUIFTEAMB-1881
Change-Id: I38adbf4dd1caf20d28e59ebb2240dc69c0c563e0