[INTERNAL] Infra: disallow the use of jQuery as a global
To better keep track of what modules still require jQuery, jQuery should
not be used via its global name but only via dependencies.

As not all code in openui5 has been cleaned up yet, the use is only
forbidden for certain libs and only for the public resources as well as
for the demokit samples and the qunit tests.

Change-Id: I517e3fc02f4c1f78b73c2b0309df41be06806aac