[INTERNAL] sap.ui.fl: Introduce new apply states for failed and successful changes
Currently the apply state of all changes is set to finished, regardless
of whether the change handler threw an error during the apply process or
not. This is a problem when relying on revert data, which might not have
been set in time in such cases.
The new change states APPLY_SUCCESSFUL and APPLY_FAILED allow to better
distinguish changes and act accordingly.

BCP: 2280062419
Change-Id: I3028909e267f0a68ea2bafaa76433ed3e0cc1d4f