[INTERNAL] mdc.Table: fix moveColumn
The moveColumn action is not sufficient for the ResponsiveTable case. The "order" of the sap.m.Column should be updated for the correct autoPopinMode calculations.
To fix the issue, the moveColumn updates the mdc.Table's column aggregation and then this updated aggregation column index is utilized to update the order for the inner sap.m.Column controls.
BCP: 2280039447

Change-Id: I385f41020ce25a00c0dfa7bb8e02fcb49e062083