[INTERNAL] sap.uxap.ObjectPageLayout: Fixed section selection mismatch
Problem:
The name of the section is not updated in the HierarchicalSelect upon
scroll on mobile size after initially picking a section different from
the selected one from the dropdown.

Root cause:
The updates of the selected item (section) in the HierarchicalSelect also
depend on correct comparison of the currently scrolled section
and the destination section.
However, when "_updateSelectionOnScroll" function compares the id of the
closest scrolled section and the id of the destination section, it
does not take into account the granularity of the ids.
So in the problematic case it compared a sectionId with a
subSectionId and found mismatch without checking if the subsection
is part of the section.

Solution:
Obtain the sectionId corresponding to the closest subSectionId in order
to check if the closest section is the destination section.

JIRA: BGSOFUIPIRIN-5426

Change-Id: I49fd8aa6b993a69138089471639c02c6de19bec6
(cherry picked from commit a27581eb3d7c0330878cd6e929cf73c4592bf387)

Change-Id: I50d190b5df90db6dd0a33203fb8036e618effc1d
BCP: 2280106171
CR-Id: 002075125900001539332022