[FIX] sap.f.DynamicPage: Wizard button on mobile correctly positioned
Problem: On mobile device the Step button of the Wizard is not correctly
positioned, when it is placed in DynamicPage.

Solution: The problem comes from the will-change CSS property of the
scrollable container of the DynamicPage. When used with "transform"
value, the stacking context does not allow the Step button to calculate
its correct position. Now the property value is changed to
"scroll-position", so that the perfromance optimization for scrolling is
not lost, but the Step button of the Wizard is positioned correctly.

BCP: 2280067783
Change-Id: Icebaad4b076d6356bf48f2538fb6622fcced1c5a