[INTERNAL] sap.ui.webc.* - do not set initial properties
The problem: "ui5-textarea" has "null" as the default value for the
"maxlength" property (meaning no restriction to the number of chars
the user can type). However, when OpenUI5 sets "null" to "maxlength" in
the wrapper control, it gets transformed to "0", which makes the control
unusable with default settings. In order to type, one must explicitly
set "maxlength" to a positive integer.

In UI5 Web Components "null" has the meaning of "not set" property.
However, in OpenUI5 "null" cannot be set for primitive types - setting
"null" enforces the default value for the type.

This change is the first part of the fix: do not set "null" as the
default value of properties in OpenUI5, but is also a broader
optimization: do not set attributes for initial properties at all.

The second part of the fix will be removing "defaultValue: null"
whenever set, which will be in a separate commit.

Note: It is possible in the future to have use cases where we want
different default values for properties in OpenUI5, compared to UI5
Web Components. If that is the case, we cannot skip setting attributes
for initial properties any more. If such cases arise, there will be a
new metadata setting in "WebComponentProperty", instructing OpenUI5 to
set an attribute for that particular property, even if initial.

Change-Id: Ia8490976d4147d34acce131f5167ad243411557c