[INTERNAL] sap.m.DateTimePicker: Improve the value parser
Problem:
The control uses the binding type formatter when parsing
the value. It does not take the control's timezone into account.

Solution:
The control always uses the timezone when parsing, regardless of
the binding type.

Problem 2:
The timezone setter updates the value and dateValue, even if
the timezone is the same. Which results in more unwanted changes
triggered by the binding.

Solution 2:
The timezone setter does not update the value when the timezone
is the same.

Change-Id: I68cf025a8e80dcba953e921dc3b476856d1275f1