[FIX] sap.m.ViewSettingsDialog: selectedSortItem value is retained
Problem:
When selectedSortItem is set before there are items, it didn't store
the value. Thus when the items came, the selection cannot be determined.

Solution:
The behavior does not depend on settings order anymore. The
selectedSortItem association retains its value and helps the selection
when the items come.

The same applies for selectedGroupItem.

BCP: 002075129400000397612022
Change-Id: I87e5fe17274bfeb0859b2dedc27226532d6ec7a2
CR-Id: 002075125900000198632022