[INTERNAL] sap.m.Button: focus is working on Safari in Horizon theme
Until now the focus in Horizon was set by offset. The issue was that
the border rounding was not applied to it in Safari.

The focus is now dislayed as before element, which is preffered instead
of after element, because the after element is used for displaying the
arrow of the MenuButton.

JIRA: BGSOFUIBALKAN-6108
Change-Id: I4ab9595250fb8bc8fc1919d47bce3c9a3eba9348