[FIX] sap.m.ViewSettingsDialog: selectedSortItem value is retained
Problem:
When selectedSortItem is set before there are items, it didn't store
the value. Thus when the items came, the selection cannot be determined.

Solution:
The behavior does not depend on settings order anymore. The selectedSortItem
association retains its value and helps the selection when the items come.

The same applies for selectedGroupItem.

BCP: 002075129400000397612022
Change-Id: I5a2817be2364d2c36a2a4d64983e87059f5d4570