[INTERNAL] sap.ui.webc.common.WebcomponentRenderer - unmanaged properties fixed
Background: WebComponentRenderer preserves all properties (such as
component private properties) that are not controlled by UI5, so that
they don't disappear between renderings.

However, if a property was mapped to an attribute with a different name
(f.e. enabled => disabled), the mapped to attribute was not properly
recognized as controlled by UI5, and was thus always preserved.

The bug was that when we call setEnabled(true), the render manager
would always preserve the "disabled" attribute, as it did not find it
among the control's properties. This is fixed by checking not only if
it is among the properties, but also if any property maps to it.

Change-Id: Ie50107dc412cf07b147b083e45b5fbc118371409