[FIX] Table: Flickering of horizontal scrollbar when in a dialog
The sap.m.Dialog triggers animations when opening and resizing. The
table reacts to this (onTableResize), and updates the horizontal
scrollbar visibility. The changes the height of the table, and the
dialog reacts to this by adjusting its height (with animation).
Due to rounding errors if zoomed in Chrome, this caused an infinite
loop, where the scrollbar is alternately shown and hidden.

Change-Id: I0b574bd52b8bff38627beb92afb7ba1ff0bb5074
BCP: 002075129400001311102022
CR-Id: 002075125800001416352022