[INTERNAL] sap.m.DynamicDateRange: visual test page has language set
Issue: The language used while developing the visual tests and the
one fromthe machine, where the visual tests are run can be different.
This causes usage of different display format in the DDR, which can
ruin the test case. Example is the rendering of AP/PM buttons which
adds additional tab step when applying the chosen value.

Solution: the language of the test page is now explicitly set.
Change-Id: I54b94c6d25bf1378a0530dd14d42b41e84360054