[INTERNAL] *FilterBar: DefaultProviderRegistry integration into filterbars
The DefaultProviderRegistry provided via the p13n Engine
allows to set a default persistence behavior.

The Adaptation & vh.FilterBar are now always registered
for transient persistence, e.g. no changes are going to be stored in the variant.

Changes are now also applied on the FilterBar derivations during runtime,
the difference to the regular approach is that the changes are discared
at the end of the session and applied through a "Dark" VariantManagement
created by the PersistenceProvider which is utilized by the DefaultProviderRegistry.t

Change-Id: I660dcd1f40b1d7c2dda0f64a46eac93b170faf4e