[INTERNAL] sap.m.PlanningCalendar: Align to timezone configuration
We are aligning the SPC and PC to a new time zone that is set as configuration feature.
This is done by fixing the time strip headers in PC and the interval headers in
in SPC to be static in all the time zones, then we align the start/end date,
of the appointments and the 'now' marker to the timezone from the core
configuration. This is done by using the internal util 'TimezoneUtil'.

Change-Id: Iaa142434f0c347c32de21fb1df01ee1f73e56386