[FIX] sap.m.TimePicker: The autocompleted values update the model once
The new value is updated once via the value base setter and once
via the onChange handler. The strings should be equal, and the model
should be updated once.
This change ensures that the auto-completed value is equal to the full
value, thus the model is updated only once.

BCP: 002075129400000146402022
Change-Id: I11eda9b39942484e689656f5f7c06dbf7931424f