[INTERNAL] mdc.Table: fix moveColumn
This is a follow-up of change I385f41020ce25a00c0dfa7bb8e02fcb49e062083.
In the above fix, we update the "order" of the inner sap.m.Column for
the correct calculations of the autoPopinBehavior. But just updating the
order was not enough, the inner responsive table's column and template
aggregations must also be updated, otherwise it lead to wrong column
header action sort behavior.
When sorting a column after moveColumn has been triggered would lead to
incorrect sorting, which is based on the column index.
Details can be found in the reported issue below.
We now update the column and template index in addition to the column
order for the responsive table.
BCP: 2270010094

Change-Id: I5b90607414239182d7bd450e93d158623279a7e8
CR-Id: 002075125900000174372022