[INTERNAL] sap.ui.webc.common.WebComponent - public methods and getters results converted
Until now only public events' parameters were processed and DOM
references were replaced with their OpenUI5 controls counterparts
(whenever appropriate).

With this change, the functionality is extended to both public methods
and public getters.

Example - the following call:

oAvatarGroup.getHiddenItems()
(which is internally proxied to the "hiddenItems" getter)

used to return an array of DOM elements (the avatar instances).
Now, it returns an array of OpenUI5 controls, which would be expected
by developers.

The same goes for public methods, which return web component instances
or arrays/objects of such.

Change-Id: Ia8d5d6467cdd7546bb1d2cff2ace31d23c8eddf2