[INTERNAL] Table: fix sticky info toolbar styles
In the older high contrast themes (sap_hcb & Belize) had rounded borders
for the info toollbar which required additional padding in the sticky
case. With the Fiori 3 and Horizon high contrast themes, the info
toolbar does not have this rounded borders and hence the addition 2px
padding is not required.
The 2px in the newer themes also created a gap, where the item content
was visible when scrolled.
This 2px padding is now removed on the info toolbar for the sticky case.
BCP: 2270016726

Change-Id: I21614664145d0ec1003f0a29e0e0e141db2b9534