[INTERNAL] mdc.Table: fix toolbar style issue
The toolbar style explicitly relied if the "type" property is defined as a string or not. This caused in correct toolbar style being applied to the ResponsiveTable type since "Table" is the default.
The Toolbar style were also not updated if the table type is changed at runtime.
The above issues are now fixed and the respective styles are applied to the toolbar.
BCP: 2270016987

Change-Id: I315b03cb6682f20a38ea9ecd2eefa6b3156b01e3