[INTERNAL] sap.ui.webc.common.WebComponent: state sync invalidation skipped
Background:
Each wrapper listens for the invalidation event of its respective web
component in order to synchronize its managed object state whenever
component properties change due to user interaction. For example, when
the user interacts with a checkbox, the web component fires its
invalidation event and the wrapper updates its "checked" property.

The problem:
We currently use "oPropData.set" to update the managed object state.
This in turn triggers an invalidation, as it directly calls the mutator
(for example "setChecked"). This leads to an invalidation, which:
 1) is unnecessary, because we already know that the web component
 instance is up to date - we only need to update the managed object
 state
 2) is problematic, because this invalidation may lead to focus loss.

 The solution:
 Instead of using "oPropData.set", directly call "setProperty" and
 prevent this invalidation.

Change-Id: I6ce04099120c9a60fd62839742e82622e513c0f7