[INTERNAL] Patcher.js: only set textContent, if the element has children
The "close" method of "Patcher.js" empties the content of the HTML
element by setting the "textContent" property.

Setting "textContent" has the side effect of making empty custom
elements lose focus in certain scenarios, even though the focus is in
the shadow root and should not be affected by "textContent". The problem
is overcome if setting textContent to "" is skipped for empty elements.

Change-Id: I002b0a987ff0182fb3e4b45315c89334a6ee6201