[INTERNAL] Table: fix sticky info toolbar styles
In the older high contrast themes (sap_hcb & Belize) had rounded borders
for the info toolbar which required additional padding in the sticky
case. With the Fiori 3 high contrast themes, the info
toolbar does not have this rounded borders and hence the addition 2px
padding is not required.
The 2px in the newer themes also created a gap, where the item content
was visible when scrolled.
This 2px padding is now removed on the info toolbar for the sticky case.
BCP: 2270016726

Change-Id: Ia5f31940b345d7e67059b7d1815715aeb20c121d
CR-Id: 002075125900000389592022